+++ services/avs/mgmt/2021-12-01/avs/CHANGELOG.md
@@ -1,2 +1,16 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Signature Changes
+
+#### Funcs
+
+1. LocationsClient.CheckTrialAvailability
+	- Params
+		- From: context.Context, string
+		- To: context.Context, string, string
+1. LocationsClient.CheckTrialAvailabilityPreparer
+	- Params
+		- From: context.Context, string
+		- To: context.Context, string, string